backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for smart_writing_coach"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "fastapi>=0.135.1",
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "uvicorn>=0.41.0",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"
